+++ R/save_diversity_maps_mw.R
@@ -4,31 +4,25 @@
 #' @param ab_div_metrics list produced from get_raster_diversity
 #' @param alpha_metrics list. alpha diversity metrics: richness, shannon, simpson, hill
 #' @param Hill_order numeric. Hill order
-#' @param Beta_info list. 
-#' @param beta_metrics list beta diversity metrics: 
+#' @param Beta_info list.
+#' @param beta_metrics list beta diversity metrics:
 #' @param fd_metrics character. list of functional metrics
 #' @param input_rast list. path for input rasters
 #' @param output_dir character.
 #' @param output_raster_name character.
 #' @param window_size numeric. window size for square plots
 #' @param filetype character. GDAL driver
+#' @param buffer_mw boolean. should buffer be applied on moving windows?
 #'
 #' @return diversity_maps character. path for diversity maps
 #' @importFrom terra aggregate writeRaster values names
 #' @export
 
-save_diversity_maps_mw <- function(input_raster_path,
-                                   ab_div_metrics,
-                                   alpha_metrics = 'shannon',
-                                   Hill_order = 1,
-                                   Beta_info = NULL, 
-                                   beta_metrics = 'bray',
-                                   fd_metrics = NULL,
-                                   input_rast,
-                                   output_dir,
-                                   output_raster_name = NULL,
-                                   window_size,
-                                   filetype = 'GTiff'){
+save_diversity_maps_mw <- function(
+    input_raster_path, ab_div_metrics, alpha_metrics = 'shannon', Hill_order = 1,
+    Beta_info = NULL, beta_metrics = 'bray', fd_metrics = NULL, input_rast,
+    output_dir, output_raster_name = NULL, window_size, filetype = 'GTiff',
+    buffer_mw = TRUE){
   # save alpha diversity indices
 
   # rast_out <- terra::rast(input_raster_path[[1]])
@@ -52,8 +46,9 @@ save_diversity_maps_mw <- function(input_raster_path,
     terra::values(template_rast[[1]]) <- ab_div_metrics[[idx]]$mean
     names(template_rast[[1]]) <- paste('mean', idx)
     # crop borders to avoid artifacts
-    template_rast <- crop_border_tile(rast_obj = template_rast,
-                                      window_size = window_size)
+    if (buffer_mw)
+      template_rast <- crop_border_tile(rast_obj = template_rast,
+                                        window_size = window_size)
     # define output raster name
     if (is.null(output_raster_name[[idx]]))
       output_raster <- file.path(output_dir, paste0(idx2, '_mean'))
@@ -72,8 +67,9 @@ save_diversity_maps_mw <- function(input_raster_path,
     terra::values(template_rast[[1]]) <- ab_div_metrics[[idx]]$sd
     names(template_rast[[1]]) <- paste('sd', idx)
     # crop borders to avoid artifacts
-    template_rast <- crop_border_tile(rast_obj = template_rast,
-                                      window_size = window_size)
+    if (buffer_mw)
+      template_rast <- crop_border_tile(rast_obj = template_rast,
+                                        window_size = window_size)
     # define output raster name
     if (is.null(output_raster_name[[idx]]))
       output_raster <- file.path(output_dir, paste0(idx2, '_sd'))
@@ -95,8 +91,9 @@ save_diversity_maps_mw <- function(input_raster_path,
     terra::values(template_rast[[1]]) <- ab_div_metrics[[idx]]
     names(template_rast[[1]]) <- idx
     # crop borders to avoid artifacts
-    template_rast <- crop_border_tile(rast_obj = template_rast,
-                                      window_size = window_size)
+    if (buffer_mw)
+      template_rast <- crop_border_tile(rast_obj = template_rast,
+                                        window_size = window_size)
     # define output raster name
     if (is.null(output_raster_name[[idx]]))
       output_raster <- file.path(output_dir, idx)
@@ -123,9 +120,10 @@ save_diversity_maps_mw <- function(input_raster_path,
         terra::values(template_rast[[nbPC]]) <- PCoA[[nbPC]]
       names(template_rast) <- paste0('PCoA#',seq(1,dimPCoA))
       # crop borders to avoid artifacts
-      template_rast <- crop_border_tile(rast_obj = template_rast,
-                                        window_size = window_size)
-      
+      if (buffer_mw)
+        template_rast <- crop_border_tile(rast_obj = template_rast,
+                                          window_size = window_size)
+
       if (is.null(output_raster_name[[beta]]))
         output_raster <- file.path(output_dir, beta)
       if (!is.null(output_raster_name[[beta]]))
@@ -136,7 +134,7 @@ save_diversity_maps_mw <- function(input_raster_path,
                          filetype = filetype, overwrite = TRUE,
                          gdal = c("COMPRESS=LZW"))
       diversity_maps[[beta]] <- output_raster
-      
+
     }
   }
   return(diversity_maps)